depression-detector-backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
idna==3.4
pymongo==3.12.3
pytz==2022.6
requests==2.28.1
sqlparse==0.2.4
urllib3==1.26.13
